Work around dl_phdr_info::dlpi_addr value in CheriABI static linking
We have to work around the position dependent linking case where dlpi_addr
will contain just the binary range (and can't a be massively out of bounds cap
with a zero vaddr due to Cheri128 constaints). In that case phdr->p_vaddr
will be within the bounds of pinfo->dlpi_addr so we just set the address
to match the vaddr.