[libunwind] Don't skip unwinding last frame in call stack
Having the CFA be exactly at the capability top is expected at the top
of the call stack on secondary threads, so fix this off-by-one error
such that those frames are not skipped, only ones where the CFA is
strictly out-of-bounds.

This is a targeted fix solely for DwarfInstructions::getCFA; it is
highly likely that some of the calls to assert_pointer_in_bounds are
overly strict. It's also likely some are overly weak since, if in
bounds, they probably want more than a single byte, but that's less of
a concern, since these are just to aid in catching nonsense states
early rather than crashing in a confusing manner later.